spin to win for earth control

+++ src/main/java/com/crowsofwar/avatar/client/particles/oldsystem/AvatarParticle.java
@@ -90,14 +90,14 @@ public abstract class AvatarParticle extends Particle {
 		int j = i >> 16 & 65535;
 		int k = i & 65535;
 		Vec3d[] avec3d = new Vec3d[]{
-				new Vec3d((double) (-rotationX * f4 - rotationXY * f4), (double) (-rotationZ * f4),
-						(double) (-rotationYZ * f4 - rotationXZ * f4)),
-				new Vec3d((double) (-rotationX * f4 + rotationXY * f4), (double) (rotationZ * f4),
-						(double) (-rotationYZ * f4 + rotationXZ * f4)),
-				new Vec3d((double) (rotationX * f4 + rotationXY * f4), (double) (rotationZ * f4),
-						(double) (rotationYZ * f4 + rotationXZ * f4)),
-				new Vec3d((double) (rotationX * f4 - rotationXY * f4), (double) (-rotationZ * f4),
-						(double) (rotationYZ * f4 - rotationXZ * f4))};
+				new Vec3d(-rotationX * f4 - rotationXY * f4, -rotationZ * f4,
+						-rotationYZ * f4 - rotationXZ * f4),
+				new Vec3d(-rotationX * f4 + rotationXY * f4, rotationZ * f4,
+						-rotationYZ * f4 + rotationXZ * f4),
+				new Vec3d(rotationX * f4 + rotationXY * f4, rotationZ * f4,
+						rotationYZ * f4 + rotationXZ * f4),
+				new Vec3d(rotationX * f4 - rotationXY * f4, -rotationZ * f4,
+						rotationYZ * f4 - rotationXZ * f4)};
 
 		if (this.particleAngle != 0.0F) {
 			float f8 = this.particleAngle + (this.particleAngle - this.prevParticleAngle) * partialTicks;
@@ -105,29 +105,29 @@ public abstract class AvatarParticle extends Particle {
 			float f10 = MathHelper.sin(f8 * 0.5F) * (float) cameraViewDir.x;
 			float f11 = MathHelper.sin(f8 * 0.5F) * (float) cameraViewDir.y;
 			float f12 = MathHelper.sin(f8 * 0.5F) * (float) cameraViewDir.z;
-			Vec3d vec3d = new Vec3d((double) f10, (double) f11, (double) f12);
+			Vec3d vec3d = new Vec3d(f10, f11, f12);
 
 			for (int l = 0; l < 4; ++l) {
 				avec3d[l] = vec3d.scale(2.0D * avec3d[l].dotProduct(vec3d))
 						.add(avec3d[l].scale((double) (f9 * f9) - vec3d.dotProduct(vec3d)))
-						.add(vec3d.crossProduct(avec3d[l]).scale((double) (2.0F * f9)));
+						.add(vec3d.crossProduct(avec3d[l]).scale(2.0F * f9));
 			}
 		}
 
 		vb.pos((double) f5 + avec3d[0].x, (double) f6 + avec3d[0].y, (double) f7 + avec3d[0].z)
-				.tex((double) maxU, (double) maxV)
+				.tex(maxU, maxV)
 				.color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha)
 				.lightmap(j, k).endVertex();
 		vb.pos((double) f5 + avec3d[1].x, (double) f6 + avec3d[1].y, (double) f7 + avec3d[1].z)
-				.tex((double) maxU, (double) minV)
+				.tex(maxU, minV)
 				.color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha)
 				.lightmap(j, k).endVertex();
 		vb.pos((double) f5 + avec3d[2].x, (double) f6 + avec3d[2].y, (double) f7 + avec3d[2].z)
-				.tex((double) minU, (double) minV)
+				.tex(minU, minV)
 				.color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha)
 				.lightmap(j, k).endVertex();
 		vb.pos((double) f5 + avec3d[3].x, (double) f6 + avec3d[3].y, (double) f7 + avec3d[3].z)
-				.tex((double) minU, (double) maxV)
+				.tex(minU, maxV)
 				.color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha)
 				.lightmap(j, k).endVertex();
